requirements.txt
eth-brownie==1.12.4
slither-analyzer==0.8.1